# test: improve language service tests performance