@alan-agius4 2019-05-21
On codefresh CI

Before
```
//packages/language-service/test:test                                    PASSED in 297.9s 
```

With the change
```
//packages/language-service/test:test                                    PASSED in 111.6s 
```

That shaves off approx 3mins

//cc @filipesilva