@alan-agius4 2019-05-24
@IgorMinar, no it’s should be master & patch.


Thanks for highlighting it.