@matsko 2019-05-24
@alan-agius4 @IgorMinar I merged this only to master ... let me know if this was the correct thing to do.