@matsko 2019-05-25
It doesn't apply cleanly. Can you create another PR based on the `patch` branch and I will merge it in?